+++ packages/scripts/scripts/lint-style.js
@@ -17,7 +17,8 @@ const {
 
 const args = getCliArgs();
 
-const hasStylelintConfig = hasCliArg( '--config' ) ||
+// See: https://github.com/stylelint/stylelint/blob/master/docs/user-guide/configuration.md#loading-the-configuration-object.
+const hasLintConfig = hasCliArg( '--config' ) ||
 	hasProjectFile( '.stylelintrc' ) ||
 	hasProjectFile( '.stylelintrc.js' ) ||
 	hasProjectFile( '.stylelintrc.json' ) ||
@@ -26,13 +27,21 @@ const hasStylelintConfig = hasCliArg( '--config' ) ||
 	hasProjectFile( '.stylelint.config.js' ) ||
 	hasPackageProp( 'stylelint' );
 
-const config = ! hasStylelintConfig ?
+const config = ! hasLintConfig ?
 	[ '--config', fromConfigRoot( '.stylelintrc.json' ) ] :
+	[];
+
+// See: https://github.com/stylelint/stylelint/blob/master/docs/user-guide/configuration.md#stylelintignore.
+const hasIgnoredFiles = hasCliArg( '--ignore-path' ) ||
+	hasProjectFile( '.stylelintignore' );
+
+const defaultIgnoreArgs = ! hasIgnoredFiles ?
+	[ '--ignore-path', fromConfigRoot( '.stylelintignore' ) ] :
 	[];
 
 const result = spawn(
 	resolveBin( 'stylelint' ),
-	[ ...config, ...args ],
+	[ ...config, ...defaultIgnoreArgs, ...args ],
 	{ stdio: 'inherit' }
 );